refactor: split CliAgentListView dialogs and add composables
- 提取 CliAgentCreateDialog 和 CliAgentEditDialog
- 创建 useSpinner 和 useAgentPolling composables
- CliAgentDetailView 硬编码颜色替换为 SCSS 变量
